travis, build, consensus: drop support for Go 1.7
